Move example from docstring to dedicated file.
